requirements.txt
Django==5.2.6
asgiref==3.9.1
sqlparse==0.5.3
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.0
PyJWT==2.9.0
django-cors-headers==4.9.0
python-decouple==3.8
pillow==11.3.0
requests==2.32.4
certifi==2025.7.14
charset-normalizer==3.4.2
idna==3.10
urllib3==2.5.0
korean-lunar-calendar==0.3.1
google-generativeai==0.8.5
google-api-core==2.28.1
google-auth==2.43.0
protobuf==5.29.5
grpcio==1.76.0
pytz==2025.2
tzdata==2025.2
colorama==0.4.6
python-dotenv==1.0.1
psycopg2-binary==2.9.9
